Moved the stories object to the librarian service.
This was done in order to simplify running the app for
non-professionals. In this way, users don't have to run
a server in order to make the app work. The server version
still exists in the server-req-version branch (and will be
updated with a server-side file to post updates to the JSON
file).